feat: consolidate settings into single page
