+++ package.json
@@ -11,16 +11,17 @@
   },
   "dependencies": {
     "@dnd-kit/core": "^6.0.6",
-    "@dnd-kit/sortable": "^7.0.2",
     "@dnd-kit/modifiers": "^6.0.1",
+    "@dnd-kit/sortable": "^7.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "zustand": "^4.4.1",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "zustand": "^4.4.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
+    "@types/uuid": "^10.0.0",
     "typescript": "^5.2.2",
     "vite": "^5.0.0",
     "vitest": "^1.1.0"
