+++ src/components/NurseCard.tsx
@@ -5,7 +5,6 @@ import { Nurse } from '../types';
 import { displayName } from '../utils/name';
 import { offAtLabel, shouldShowOffAt } from '../utils/time';
 import { useStore } from '../store';
-import NurseMenu from './NurseMenu';
 
 interface Props {
   nurse: Nurse;
@@ -27,7 +26,6 @@ const NurseCard: React.FC<Props> = ({ nurse, zoneId, index }) => {
 
   const [editingRF, setEditingRF] = useState(false);
   const [rfValue, setRfValue] = useState(nurse.rfNumber || '');
-  const [menuOpen, setMenuOpen] = useState(false);
 
   const saveRf = () => {
     updateNurse(nurse.id, { rfNumber: rfValue });
@@ -90,23 +88,21 @@ const NurseCard: React.FC<Props> = ({ nurse, zoneId, index }) => {
         ) : (
           <span onClick={() => setEditingRF(true)}>{nurse.rfNumber || 'RF #'}</span>
         )}
-        <button className="menu-btn" onClick={() => setMenuOpen(true)}>
+        <button
+          className="menu-btn"
+          onClick={(e) => {
+            e.stopPropagation();
+            useStore.getState().setUi({
+              contextMenu: {
+                nurseId: nurse.id,
+                anchor: e.currentTarget.getBoundingClientRect(),
+              },
+            });
+          }}
+        >
           ⋮
         </button>
       </div>
-      {menuOpen && (
-        <NurseMenu
-          nurse={nurse}
-          onClose={() => setMenuOpen(false)}
-          onChangeHospitalId={(val) => {
-            try {
-              updateNurse(nurse.id, { hospitalId: val || undefined });
-            } catch (err) {
-              alert((err as Error).message);
-            }
-          }}
-        />
-      )}
     </div>
   );
 };
